Fix: stack score가 config가 아닌 고정값 0.1로 적용되는 버그

+++ src/main/java/com/doubledeltas/minecollector/config/chapter/ScoringChapter.java
@@ -13,7 +13,7 @@ public class ScoringChapter {
 
     private boolean     stackEnabled;
     private int         stackMultiple;
-    private float       stackScore;
+    private BigDecimal  stackScore;
 
     private boolean     advancementEnabled;
     private Map<AdvancementDisplayType, BigDecimal> advancementScores;
